Create test configurations separately
